Dropped EventStoreDBRepository in favour to extension methods
